+++ RELEASE-NOTES.md
@@ -3,4 +3,5 @@
 * #1 OpenTelemetry bootstrap
 * #2 Configurable SdkTracerProvider
 * #3 Injectable OTel Resource with Bootique application name
+* #4 Configurable SdkMeterProvider